requirements.txt
alpaca-py>=0.8.0
yfinance>=0.2.0
pandas>=1.5.0
numpy>=1.23.0
requests>=2.28.0
schedule>=1.2.0
pytz>=2022.1
